Refactor: Update file structure to include `src` folder for each service or library. * Refactor: Remove root `package.json`, `package-lock.json`, `babel.config.js`, and `.eslintrc.json`, `jest.config.ts`, and `tsconfig.json` files. * Config: Add `.eslintrc.json` file to all node services and libraries.
