--- babel.config.js
@@ -1,4 +1,0 @@
-export const presets = [
-	["@babel/preset-env", { targets: { node: "current" } }],
-	"@babel/preset-typescript",
-];
