--- package.json
@@ -1,50 +1,0 @@
-{
-  "name": "backend-services",
-  "version": "0.0.1",
-  "description": "Mono repo containing https://leetcoderoulette.com backend microservices.",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --passWithNoTests",
-    "test:ci": "jest --coverage"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/leetcode-roulette/backend-services.git"
-  },
-  "keywords": [
-    "Microservices",
-    "Leetcode",
-    "Roulette",
-    "Leetcode",
-    "API",
-    "RestAPI",
-    "Rest",
-    "Microservice",
-    "Architecture"
-  ],
-  "author": "Leetcode Roulette",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/leetcode-roulette/backend-services/issues"
-  },
-  "homepage": "https://github.com/leetcode-roulette/backend-services#readme",
-  "devDependencies": {
-    "@babel/core": "^7.19.1",
-    "@babel/preset-env": "^7.19.1",
-    "@babel/preset-typescript": "^7.18.6",
-    "@types/jest": "^29.0.3",
-    "@types/node": "^18.7.18",
-    "@types/supertest": "^2.0.12",
-    "@typescript-eslint/eslint-plugin": "^5.42.0",
-    "@typescript-eslint/parser": "^5.42.0",
-    "eslint": "^8.26.0",
-    "jest": "^29.2.2",
-    "supertest": "^6.3.1",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.4"
-  },
-  "dependencies": {
-    "dotenv": "^16.0.3"
-  }
-}
