+++ src/libs/leetcodeAPI/src/interfaces/leetcodeAPI.ts
