+++ src/libs/leetcodeAPI/tsconfig.json
@@ -1,3 +1,19 @@
 {
-  "extends": "../../../tsconfig.json"
+	"compilerOptions": {
+		"strict": true,
+		"declaration": true,
+		"checkJs": true,
+		"allowJs": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+		"declarationMap": true,
+		"baseUrl": "./src",
+		"outDir": "dist",
+		"target": "ES2022",
+		"module": "CommonJS",
+		"noImplicitAny": true,
+    "resolveJsonModule": true
+	},
+	"include": ["./src"],
+	"exclude": ["node_modules"]
 }
