+++ src/services/tags/jest.config.ts
@@ -1,3 +1,8 @@
+import { pathsToModuleNameMapper } from "ts-jest";
+import config from "./tsconfig.json";
+
+const moduleNameMapper = pathsToModuleNameMapper(config.compilerOptions.paths);
+
 export default {
 	roots: ["<rootDir>"],
 	preset: "ts-jest",
@@ -12,5 +17,7 @@ export default {
 		"**/tests/**/*.spec.ts",
 		"**/tests/**/*.test.ts",
 	],
-	testEnvironment: "node"
+	testEnvironment: "node",
+	moduleNameMapper,
+	modulePaths: [config.compilerOptions.baseUrl]
 };
