+++ src/services/tags/package.json
@@ -5,7 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "jest --passWithNoTests",
-    "dev": "nodemon index.ts"
+    "dev": "nodemon src/index.ts",
+    "build": "tsc",
+    "start": ""
   },
   "keywords": [
     "Tags",
@@ -31,20 +33,20 @@
     "winston": "^3.8.2"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.14",
-    "nodemon": "^2.0.20",
-    "supertest": "^6.3.1",
     "@babel/core": "^7.19.1",
     "@babel/preset-env": "^7.19.1",
     "@babel/preset-typescript": "^7.18.6",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.14",
     "@types/jest": "^29.0.3",
     "@types/node": "^18.7.18",
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^5.42.0",
     "@typescript-eslint/parser": "^5.42.0",
-    "eslint": "^8.26.0",
+    "eslint": "^8.27.0",
     "jest": "^29.2.2",
+    "nodemon": "^2.0.20",
+    "supertest": "^6.3.1",
     "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
