--- tsconfig.json
@@ -1,24 +1,0 @@
-{
-	"compilerOptions": {
-		"strict": true,
-		"declaration": true,
-		"checkJs": true,
-		"allowJs": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"declarationMap": true,
-		"paths": {
-			"@libs/*": ["libs/*"],
-			"@services/*": ["services/*"],
-			"@shared/*": ["shared/*"]
-		},
-		"baseUrl": "./src",
-		"outDir": "dist",
-		"target": "ES2022",
-		"module": "CommonJS",
-		"noImplicitAny": true,
-    "resolveJsonModule": true
-	},
-	"include": ["./src"],
-	"exclude": ["node_modules"]
-}
